+++ shape/actor.go
@@ -30,7 +30,7 @@ func (a *Actor) Width() int            { return a.rad() * 4 }
 func (a *Actor) Height() int           { return a.height }
 func (a *Actor) SetHeight(h int)       { a.height = h }
 func (a *Actor) rad() int              { return a.height / 6 }
-func (a *Actor) Direction() Direction  { return LR }
+func (a *Actor) Direction() Direction  { return RightDir }
 func (a *Actor) SetClass(class string) { a.class = class }
 
 func (a *Actor) WriteSvg(out io.Writer) error {